requirements.txt
icalendar>=5.0.0
python-dateutil>=2.8.2
pytz>=2023.3
pytest>=7.4.0
pandas>=2.0.0
openpyxl>=3.1.0
typing-extensions>=4.7.0